Add files via upload
NASA KEPLER TELESCOPE DATA SET :  Long back, NASA placed a telescope called the Kepler telescope in space. This telescope is used to measure the brightness of the stars in the far-distant galaxies. with the help of this variation in brightness which is  captured by the telescope , it can be predicted that star has a exoplanet or not.
so in this project we will clean the data and check for the missing values and incorrect data and replace that with the  appropriate data after that we will use this data set to analyze brightness pattern to predict the exoplanets.